backend/requirements.txt
openai
qdrant-client
python-dotenv
tqdm
pydantic
fastapi
uvicorn[standard]
